Clarity#1240 - Update Stamps View
